+++ src/game/tests.rs
@@ -665,6 +665,7 @@ mod unblocked_attackers;
 mod unbounded_targets;
 mod unleash;
 mod untap_caps;
+mod untap_skip_spells;
 mod untap_skips;
 mod until_end_of_combat;
 mod urza_lands;